+++ generators/generators.js
@@ -6,9 +6,11 @@ import {
   meshMap,
   meshBlockIdMap,
   generateUniqueId,
-} from "./meshState.js";
+} from "./mesh-state.js";
+import { registerTransformGenerators } from "./generators-transform.js";
+import { getFieldValue } from "./generators-utilities.js";
 
-export * from "./meshState.js";
+export * from "./mesh-state.js";
 
 function sanitizeForCode(input) {
   let s = String(input);
@@ -138,6 +140,8 @@ function emitSafeIdentifierLiteral(code) {
 }
 
 export function defineGenerators() {
+  registerTransformGenerators(javascriptGenerator);
+
   const reservedWordsWithoutName = javascriptGenerator.RESERVED_WORDS_.split(
     ",",
   )
@@ -166,16 +170,6 @@ export function defineGenerators() {
 
     return `await hide(${modelName});\n`;
   };
-
-  function getFieldValue(block, fieldName, defaultValue) {
-    return (
-      javascriptGenerator.valueToCode(
-        block,
-        fieldName,
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || defaultValue
-    );
-  }
 
   function getVariableInfo(block, fieldName) {
     const variableId = block.getFieldValue(fieldName);
@@ -222,163 +216,6 @@ export function defineGenerators() {
       ) || "false"; // Default to false if no condition is connected
 
     return `await waitUntil(() => ${condition});\n`;
-  };
-
-  javascriptGenerator.forBlock["glide_to"] = function (block) {
-    const meshVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const x =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const y =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const z =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const duration =
-      javascriptGenerator.valueToCode(
-        block,
-        "DURATION",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "1000";
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-
-    const code = `${mode === "AWAIT" ? "await " : ""}glideTo(${meshVar}, { x: ${x}, y: ${y}, z: ${z}, duration: ${duration} / 1000, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
-
-    return code;
-  };
-
-  javascriptGenerator.forBlock["glide_to_seconds"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-    const duration = getFieldValue(block, "DURATION", "0");
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-
-    const asyncWrapper = mode === "AWAIT" ? "await " : "";
-
-    return `${asyncWrapper}glideTo(${meshName}, { x: ${x}, y: ${y}, z: ${z}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
-  };
-
-  javascriptGenerator.forBlock["glide_to_object"] = function (block) {
-    const meshName1 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL1"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshName2 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL2"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const xOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "X_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const zOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const duration = getFieldValue(block, "DURATION", "0");
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-    const asyncWrapper = mode === "AWAIT" ? "await " : "";
-
-    return `${asyncWrapper}glideToObject(${meshName1}, ${meshName2}, { offsetX: ${xOffset}, offsetY: ${yOffset}, offsetZ: ${zOffset}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
-  };
-
-  javascriptGenerator.forBlock["rotate_anim"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const rotX = getFieldValue(block, "ROT_X", "0");
-    const rotY = getFieldValue(block, "ROT_Y", "0");
-    const rotZ = getFieldValue(block, "ROT_Z", "0");
-    const duration = getFieldValue(block, "DURATION", "0");
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-
-    const asyncWrapper = mode === "AWAIT" ? "await " : "";
-
-    return `${asyncWrapper}rotateAnim(${meshName}, { x: ${rotX}, y: ${rotY}, z: ${rotZ}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
-  };
-
-  javascriptGenerator.forBlock["rotate_anim_seconds"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const rotX = getFieldValue(block, "ROT_X", "0");
-    const rotY = getFieldValue(block, "ROT_Y", "0");
-    const rotZ = getFieldValue(block, "ROT_Z", "0");
-    const duration = getFieldValue(block, "DURATION", "0");
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-
-    const asyncWrapper = mode === "AWAIT" ? "await " : "";
-
-    return `${asyncWrapper}rotateAnim(${meshName}, { x: ${rotX}, y: ${rotY}, z: ${rotZ}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
-  };
-
-  javascriptGenerator.forBlock["rotate_to_object"] = function (block) {
-    const meshName1 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL1"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const meshName2 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL2"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const rotateMode = block.getFieldValue("ROTATE_MODE");
-    const apiRotateMode =
-      rotateMode === "SAME_ROTATION" ? "same_rotation" : "towards";
-    const duration = getFieldValue(block, "DURATION", "0");
-    const mode = block.getFieldValue("MODE");
-    const reverse = block.getFieldValue("REVERSE") === "TRUE";
-    const loop = block.getFieldValue("LOOP") === "TRUE";
-    const easing = block.getFieldValue("EASING");
-
-    const asyncWrapper = mode === "AWAIT" ? "await " : "";
-
-    return `${asyncWrapper}rotateToObject(${meshName1}, ${meshName2}, { mode: "${apiRotateMode}", duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
   };
 
   javascriptGenerator.forBlock["animation"] = function (block) {
@@ -656,34 +493,6 @@ export function defineGenerators() {
 
     // Return the string value as a quoted literal
     return [`"${pivotOption}"`, javascriptGenerator.ORDER_ATOMIC];
-  };
-
-  javascriptGenerator.forBlock["set_pivot"] = function (block) {
-    const meshVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const xPivot =
-      javascriptGenerator.valueToCode(
-        block,
-        "X_PIVOT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || 0;
-    const yPivot =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y_PIVOT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || 0;
-    const zPivot =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z_PIVOT",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || 0;
-
-    return `await setAnchor(${meshVar}, { xPivot: ${xPivot}, yPivot: ${yPivot}, zPivot: ${zPivot} });\n`;
   };
 
   javascriptGenerator.forBlock["start"] = function (block) {
@@ -1741,112 +1550,11 @@ export function defineGenerators() {
     return `${variableName} = newWall(${color}, ${startX}, ${startZ}, ${endX}, ${endZ}, ${yPosition}, "${wallType}", "${wallId}");\n`;
   };
 
-  javascriptGenerator.forBlock["move_by_xyz"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BLOCK_NAME"),
+  javascriptGenerator.forBlock["glide_to"] = function (block) {
+    const meshVar = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MESH_VAR"),
       Blockly.Names.NameType.VARIABLE,
     );
-
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-
-    return `await moveByVector(${modelName}, { x: ${x}, y: ${y}, z: ${z} });\n`;
-  };
-
-  javascriptGenerator.forBlock["move_by_xyz_single"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BLOCK_NAME"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const coordinate = block.getFieldValue("COORDINATE") || "x_coordinate";
-    const value = getFieldValue(block, "VALUE", "0");
-
-    switch (coordinate) {
-      case "x_coordinate":
-        return `await moveByVector(${modelName}, { x: ${value}, y: 0, z: 0 });\n`;
-      case "y_coordinate":
-        return `await moveByVector(${modelName}, { x: 0, y: ${value}, z: 0 });\n`;
-      case "z_coordinate":
-        return `await moveByVector(${modelName}, { x: 0, y: 0, z: ${value} });\n`;
-    }
-  };
-
-  javascriptGenerator.forBlock["scale"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BLOCK_NAME"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-
-    // Retrieve the origin values for x, y, and z axes
-    const xOrigin = block.getFieldValue("X_ORIGIN") || "'CENTRE'";
-    const yOrigin = block.getFieldValue("Y_ORIGIN") || "'CENTRE'";
-    const zOrigin = block.getFieldValue("Z_ORIGIN") || "'CENTRE'";
-
-    return `await scale(${modelName}, { x: ${x}, y: ${y}, z: ${z}, xOrigin: '${xOrigin}', yOrigin: '${yOrigin}', zOrigin: '${zOrigin}' });\n`;
-  };
-
-  javascriptGenerator.forBlock["resize"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BLOCK_NAME"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-
-    // Retrieve the origin values for x, y, and z axes
-    const xOrigin = block.getFieldValue("X_ORIGIN") || "'CENTRE'";
-    const yOrigin = block.getFieldValue("Y_ORIGIN") || "'CENTRE'";
-    const zOrigin = block.getFieldValue("Z_ORIGIN") || "'CENTRE'";
-
-    return `await resize(${modelName}, { width: ${x}, height: ${y}, depth: ${z}, xOrigin: '${xOrigin}', yOrigin: '${yOrigin}', zOrigin: '${zOrigin}' });\n`;
-  };
-
-  javascriptGenerator.forBlock["look_at"] = function (block) {
-    const meshName1 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL1"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshName2 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL2"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const useY = block.getFieldValue("USE_Y") === "TRUE";
-
-    //return `await lookAt(${meshName1}, ${meshName2}, ${useY});\n`;
-
-    return `await lookAt(${meshName1}, { target: ${meshName2}, useY: ${useY} });\n`;
-  };
-
-  javascriptGenerator.forBlock["move_to"] = function (block) {
-    const meshName1 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL1"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshName2 = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL2"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const useY = block.getFieldValue("USE_Y") === "TRUE";
-
-    return `await moveTo(${meshName1}, { target: ${meshName2}, useY: ${useY} });\n`;
-  };
-
-  javascriptGenerator.forBlock["rotate_to"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
     const x =
       javascriptGenerator.valueToCode(
         block,
@@ -1865,50 +1573,138 @@ export function defineGenerators() {
         "Z",
         javascriptGenerator.ORDER_ATOMIC,
       ) || "0";
+    const duration =
+      javascriptGenerator.valueToCode(
+        block,
+        "DURATION",
+        javascriptGenerator.ORDER_ATOMIC,
+      ) || "1000";
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
 
-    return `await rotateTo(${meshName}, { x: ${x}, y: ${y}, z: ${z} });\n`;
+    const code = `${mode === "AWAIT" ? "await " : ""}glideTo(${meshVar}, { x: ${x}, y: ${y}, z: ${z}, duration: ${duration} / 1000, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
+
+    return code;
   };
 
-  javascriptGenerator.forBlock["move_to_xyz"] = function (block) {
+  javascriptGenerator.forBlock["glide_to_seconds"] = function (block) {
     const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL"),
+      block.getFieldValue("MESH_VAR"),
+      Blockly.Names.NameType.VARIABLE,
+    );
+    const x = getFieldValue(block, "X", "0");
+    const y = getFieldValue(block, "Y", "0");
+    const z = getFieldValue(block, "Z", "0");
+    const duration = getFieldValue(block, "DURATION", "0");
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
+
+    const asyncWrapper = mode === "AWAIT" ? "await " : "";
+
+    return `${asyncWrapper}glideTo(${meshName}, { x: ${x}, y: ${y}, z: ${z}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
+  };
+
+  javascriptGenerator.forBlock["glide_to_object"] = function (block) {
+    const meshName1 = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MODEL1"),
       Blockly.Names.NameType.VARIABLE,
     );
 
-    const x =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const y =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const z =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-
-    const useY = block.getFieldValue("USE_Y") === "TRUE";
-
-    return `await positionAt(${meshName}, { x: ${x}, y: ${y}, z: ${z}, useY: ${useY} });\n`;
-  };
-
-  javascriptGenerator.forBlock["move_to_xyz_single"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL"),
+    const meshName2 = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MODEL2"),
       Blockly.Names.NameType.VARIABLE,
     );
+    const xOffset =
+      javascriptGenerator.valueToCode(
+        block,
+        "X_OFFSET",
+        javascriptGenerator.ORDER_ATOMIC,
+      ) || "0";
+    const yOffset =
+      javascriptGenerator.valueToCode(
+        block,
+        "Y_OFFSET",
+        javascriptGenerator.ORDER_ATOMIC,
+      ) || "0";
+    const zOffset =
+      javascriptGenerator.valueToCode(
+        block,
+        "Z_OFFSET",
+        javascriptGenerator.ORDER_ATOMIC,
+      ) || "0";
+    const duration = getFieldValue(block, "DURATION", "0");
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
+    const asyncWrapper = mode === "AWAIT" ? "await " : "";
 
-    const coordinate = block.getFieldValue("COORDINATE") || "x_coordinate";
-    const value = getFieldValue(block, "VALUE", "0");
+    return `${asyncWrapper}glideToObject(${meshName1}, ${meshName2}, { offsetX: ${xOffset}, offsetY: ${yOffset}, offsetZ: ${zOffset}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
+  };
 
-    return `await positionAtSingleCoordinate(${meshName}, "${coordinate}", ${value});\n`;
+  javascriptGenerator.forBlock["rotate_anim"] = function (block) {
+    const meshName = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MESH_VAR"),
+      Blockly.Names.NameType.VARIABLE,
+    );
+    const rotX = getFieldValue(block, "ROT_X", "0");
+    const rotY = getFieldValue(block, "ROT_Y", "0");
+    const rotZ = getFieldValue(block, "ROT_Z", "0");
+    const duration = getFieldValue(block, "DURATION", "0");
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
+
+    const asyncWrapper = mode === "AWAIT" ? "await " : "";
+
+    return `${asyncWrapper}rotateAnim(${meshName}, { x: ${rotX}, y: ${rotY}, z: ${rotZ}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
+  };
+
+  javascriptGenerator.forBlock["rotate_anim_seconds"] = function (block) {
+    const meshName = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MESH_VAR"),
+      Blockly.Names.NameType.VARIABLE,
+    );
+    const rotX = getFieldValue(block, "ROT_X", "0");
+    const rotY = getFieldValue(block, "ROT_Y", "0");
+    const rotZ = getFieldValue(block, "ROT_Z", "0");
+    const duration = getFieldValue(block, "DURATION", "0");
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
+
+    const asyncWrapper = mode === "AWAIT" ? "await " : "";
+
+    return `${asyncWrapper}rotateAnim(${meshName}, { x: ${rotX}, y: ${rotY}, z: ${rotZ}, duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
+  };
+
+  javascriptGenerator.forBlock["rotate_to_object"] = function (block) {
+    const meshName1 = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MODEL1"),
+      Blockly.Names.NameType.VARIABLE,
+    );
+    const meshName2 = javascriptGenerator.nameDB_.getName(
+      block.getFieldValue("MODEL2"),
+      Blockly.Names.NameType.VARIABLE,
+    );
+    const rotateMode = block.getFieldValue("ROTATE_MODE");
+    const apiRotateMode =
+      rotateMode === "SAME_ROTATION" ? "same_rotation" : "towards";
+    const duration = getFieldValue(block, "DURATION", "0");
+    const mode = block.getFieldValue("MODE");
+    const reverse = block.getFieldValue("REVERSE") === "TRUE";
+    const loop = block.getFieldValue("LOOP") === "TRUE";
+    const easing = block.getFieldValue("EASING");
+
+    const asyncWrapper = mode === "AWAIT" ? "await " : "";
+
+    return `${asyncWrapper}rotateToObject(${meshName1}, ${meshName2}, { mode: "${apiRotateMode}", duration: ${duration}, reverse: ${reverse}, loop: ${loop}, easing: "${easing}" });\n`;
   };
 
   javascriptGenerator.forBlock["distance_to"] = function (block) {
@@ -2013,19 +1809,6 @@ export function defineGenerators() {
     const code = `${idVar} = ${asyncMode === "AWAIT" ? "await " : ""}playSound(${meshName}, { soundName: "${soundName}", loop: ${loop}, volume: ${volumeCode}, playbackRate: ${speedCode} });\n`;
 
     return code;
-  };
-
-  javascriptGenerator.forBlock["rotate_model_xyz"] = function (block) {
-    const meshName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const x = getFieldValue(block, "X", "0");
-    const y = getFieldValue(block, "Y", "0");
-    const z = getFieldValue(block, "Z", "0");
-
-    return `await rotate(${meshName}, { x: ${x}, y: ${y}, z: ${z} });\n`;
   };
 
   javascriptGenerator.forBlock["forever"] = function (block) {
@@ -2521,35 +2304,6 @@ export function defineGenerators() {
     return `createMap("${mapName}", ${material});\n`;
   };
 
-  javascriptGenerator.forBlock["move_forward"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const speed =
-      javascriptGenerator.valueToCode(
-        block,
-        "SPEED",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const direction = block.getFieldValue("DIRECTION");
-
-    // Choose the appropriate helper function based on the direction
-    let helperFunction;
-    switch (direction) {
-      case "sideways":
-        helperFunction = "moveSideways";
-        break;
-      case "strafe":
-        helperFunction = "strafe";
-        break;
-      default:
-        helperFunction = "moveForward";
-    }
-
-    return `${helperFunction}(${modelName}, ${speed});\n`;
-  };
-
   javascriptGenerator.forBlock["control_animation_group"] = function (block) {
     const animationGroupName = javascriptGenerator.nameDB_.getName(
       block.getFieldValue("GROUP_NAME"),
@@ -2582,37 +2336,6 @@ export function defineGenerators() {
     const upForce = getFieldValue(block, "UP_FORCE", "1"); // Default up force
 
     return `up(${modelName}, ${upForce});\n`;
-  };
-
-  javascriptGenerator.forBlock["apply_force"] = function (block) {
-    // Get the name of the mesh variable
-    const mesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    // Get the force values
-    const forceX =
-      javascriptGenerator.valueToCode(
-        block,
-        "X",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const forceY =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const forceZ =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-
-    // Generate the code
-    return `applyForce(${mesh}, { forceX: ${forceX}, forceY: ${forceY}, forceZ: ${forceZ} });\n`;
   };
 
   javascriptGenerator.forBlock["touching_surface"] = function (block) {
@@ -2682,149 +2405,6 @@ export function defineGenerators() {
     return `exportMesh(${meshVar}, "${format}");\n`;
   };
 
-  javascriptGenerator.forBlock["merge_meshes"] = function (block) {
-    const resultVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("RESULT_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshList =
-      javascriptGenerator.valueToCode(
-        block,
-        "MESH_LIST",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "[]";
-
-    const meshId = "merged" + "_" + generateUniqueId();
-    meshMap[meshId] = block;
-    meshBlockIdMap[meshId] = block.id;
-
-    // Use helper function to merge the meshes
-    return `${resultVar} = await mergeMeshes("${meshId}", ${meshList});\n`;
-  };
-
-  javascriptGenerator.forBlock["subtract_meshes"] = function (block) {
-    const resultVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("RESULT_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const baseMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("BASE_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshList =
-      javascriptGenerator.valueToCode(
-        block,
-        "MESH_LIST",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "[]";
-
-    const meshId = "subtracted" + "_" + generateUniqueId();
-    meshMap[meshId] = block;
-    meshBlockIdMap[meshId] = block.id;
-
-    // Use helper function to subtract meshes from the base mesh
-    return `${resultVar} = await subtractMeshes("${meshId}", ${baseMesh}, ${meshList});\n`;
-  };
-
-  javascriptGenerator.forBlock["intersection_meshes"] = function (block) {
-    const resultVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("RESULT_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshList =
-      javascriptGenerator.valueToCode(
-        block,
-        "MESH_LIST",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "[]";
-
-    const meshId = "intersected" + "_" + generateUniqueId();
-    meshMap[meshId] = block;
-    meshBlockIdMap[meshId] = block.id;
-
-    // Use helper function to intersect the meshes
-    return `${resultVar} = await intersectMeshes("${meshId}", ${meshList});\n`;
-  };
-  javascriptGenerator.forBlock["hull_meshes"] = function (block) {
-    const resultVar = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("RESULT_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const meshList =
-      javascriptGenerator.valueToCode(
-        block,
-        "MESH_LIST",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "[]";
-
-    const meshId = "hull" + "_" + generateUniqueId();
-    meshMap[meshId] = block;
-    meshBlockIdMap[meshId] = block.id;
-
-    // Use helper function to create the hull
-    return `${resultVar} = await createHull("${meshId}", ${meshList});\n`;
-  };
-
-  javascriptGenerator.forBlock["parent"] = function (block) {
-    const parentMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("PARENT_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const childMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("CHILD_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    // Establish the parent-child relationship with offset
-    return `setParent(${parentMesh}, ${childMesh});\n`;
-  };
-
-  javascriptGenerator.forBlock["parent_child"] = function (block) {
-    const parentMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("PARENT_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const childMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("CHILD_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const xOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "X_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const zOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-
-    // Establish the parent-child relationship with offset
-    return `parentChild(${parentMesh}, ${childMesh}, ${xOffset}, ${yOffset}, ${zOffset});\n`;
-  };
-
-  javascriptGenerator.forBlock["remove_parent"] = function (block) {
-    const childMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("CHILD_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    return `removeParent(${childMesh});\n`;
-  };
-
   javascriptGenerator.forBlock["hold"] = function (block) {
     const meshToAttach = javascriptGenerator.nameDB_.getName(
       block.getFieldValue("MESH_TO_ATTACH"),
@@ -2856,126 +2436,6 @@ export function defineGenerators() {
     // Establish the hold action with offset
     return `await hold(${meshToAttach}, ${targetMesh}, ${xOffset}, ${yOffset}, ${zOffset});
         `;
-  };
-
-  javascriptGenerator.forBlock["attach"] = function (block) {
-    const meshToAttach = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_TO_ATTACH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const targetMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("TARGET_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const boneName = block.getFieldValue("BONE_NAME");
-    const xOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "X_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const zOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    // Establish the attach action with bone name and offset
-    return `await attach(${meshToAttach}, ${targetMesh}, { boneName: "${boneName}", x: ${xOffset}, y: ${yOffset}, z: ${zOffset} });
-        `;
-  };
-
-  javascriptGenerator.forBlock["drop"] = function (block) {
-    const meshToDetach = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MESH_TO_DETACH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    // Establish the drop action
-    return `await drop(${meshToDetach});
-        `;
-  };
-
-  javascriptGenerator.forBlock["follow"] = function (block) {
-    const followerMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("FOLLOWER_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const targetMesh = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("TARGET_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-    const followPosition = block.getFieldValue("FOLLOW_POSITION");
-
-    const xOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "X_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const yOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Y_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-    const zOffset =
-      javascriptGenerator.valueToCode(
-        block,
-        "Z_OFFSET",
-        javascriptGenerator.ORDER_ATOMIC,
-      ) || "0";
-
-    // Use the helper method makeFollow for following the target
-    const code = `
-                        makeFollow(${followerMesh}, ${targetMesh}, "${followPosition}", ${xOffset}, ${yOffset}, ${zOffset});
-                `;
-    return code;
-  };
-
-  javascriptGenerator.forBlock["stop_follow"] = function (block) {
-    const followerModelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("FOLLOWER_MESH"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    // Generate code to call the stopFollow helper function
-    const code = `stopFollow("${followerModelName}");\n`;
-    return code;
-  };
-
-  javascriptGenerator.forBlock["add_physics"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const physicsType = block.getFieldValue("PHYSICS_TYPE");
-
-    // Note: Ensure that the execution environment supports async/await at this level
-    return `await setPhysics(${modelName}, "${physicsType}");\n`;
-  };
-  javascriptGenerator.forBlock["add_physics_shape"] = function (block) {
-    const modelName = javascriptGenerator.nameDB_.getName(
-      block.getFieldValue("MODEL_VAR"),
-      Blockly.Names.NameType.VARIABLE,
-    );
-
-    const shapeType = block.getFieldValue("SHAPE_TYPE");
-
-    // Note: Ensure that the execution environment supports async/await at this level
-    return `await setPhysicsShape(${modelName}, "${shapeType}");\n`;
-  };
-
-  javascriptGenerator.forBlock["show_physics"] = function (block) {
-    const show = block.getFieldValue("SHOW") === "TRUE";
-    return `showPhysics(${show});\n`;
   };
 
   javascriptGenerator.forBlock["canvas_controls"] = function (block) {
